+++ src/rules/enforce-relation-types.ts
@@ -1,4 +1,4 @@
-import { AST_NODE_TYPES, ESLintUtils, TSESTree } from '@typescript-eslint/utils';
+import { AST_NODE_TYPES, ESLintUtils } from '@typescript-eslint/utils';
 import { ReportSuggestionArray } from '@typescript-eslint/utils/ts-eslint';
 import {
     findEitherDecoratorArguments,
@@ -20,6 +20,7 @@ const createRule = ESLintUtils.RuleCreator(
 );
 
 type EnforceColumnMessages =
+    | 'typescript_typeorm_relation_missing'
     | 'typescript_typeorm_relation_mismatch'
     | 'typescript_typeorm_relation_array_to_many'
     | 'typescript_typeorm_relation_nullable_by_default'
@@ -37,6 +38,8 @@ const enforceColumnTypes = createRule({
         },
         hasSuggestions: true,
         messages: {
+            typescript_typeorm_relation_missing:
+                'Relation {{ relation }} of {{ propertyName }}{{ className }} does not have an arrow function with the relation type.',
             typescript_typeorm_relation_mismatch:
                 'Type of {{ propertyName }}{{ className }} is not consistent with the TypeORM relation type {{ relation }}{{ expectedValue }}.',
             typescript_typeorm_relation_array_to_many:
@@ -66,7 +69,21 @@ const enforceColumnTypes = createRule({
                 const [relation, relArguments] = relationArguments;
                 const typeormType = convertArgumentToRelationType(relation, relArguments);
                 if (!typeormType) {
-                    return; // TODO: report error
+                    const propertyName =
+                        node.key?.type === AST_NODE_TYPES.Identifier ? node.key.name : 'property';
+                    const classObject = findParentClass(node);
+                    const className = classObject?.id ? ` in ${classObject.id.name}` : '';
+                    context.report({
+                        node,
+                        messageId: 'typescript_typeorm_relation_missing',
+                        data: {
+                            className,
+                            propertyName,
+                            relation,
+                        },
+                        loc: node.loc,
+                    });
+                    return;
                 }
 
                 if (!node.typeAnnotation) {