+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - Date column decorators `CreateDateColumn`, `UpdateDateColumn` is date by default, `DeleteDateColumn` is nullable,
     - Version column decorator `VersionColumn` is number.
 - Allow lazy relations with `Promise<Relation>` ([#10](https://github.com/daniel7grant/eslint-plugin-typeorm-typescript/issues/10))
+- Report error when the relation doesn't have an arrow function defined
 
 ### Changed